+++ src/store/index.ts
@@ -2,5 +2,6 @@ export * from "./backup";
 export * from "./database";
 export * from "./errors";
 export * from "./import-export";
+export * from "./queries";
 export * from "./schema";
 export * from "./syx-codec";